MSSBC-112: Fix message receiving when using MS Service Bus and Apache Qpid
When receiving a message from Azure Service Bus using Apache Qpid, the JMS
Connector tries to parse the JMS Message properties. One of the properties
is a java.util.UUID, and when calling message.getObjectProperty(key) it
throws an IllegalArgumentException, preventing the connector from returning
the message. This is not a fatal error, so I added this exception to the
try/catch clause that retrieves these properties.
I think you guys contemplated borderline cases like this, but for some
reason the JMS Message is not throwing a JMSException as it should.